+++ exporter/opentelemetry-exporter-jaeger-proto-grpc/src/opentelemetry/exporter/jaeger/proto/grpc/__init__.py
@@ -68,6 +68,7 @@ API
 import logging
 from os import environ
 from typing import Optional
+from deprecated import deprecated
 
 from grpc import ChannelCredentials, RpcError, insecure_channel, secure_channel
 
@@ -110,6 +111,10 @@ class JaegerExporter(SpanExporter):
         timeout: Maximum time the Jaeger exporter should wait for each batch export.
     """
 
+    @deprecated(
+        version="1.16.0",
+        reason="Since v1.35, the Jaeger supports OTLP natively. Please use the OTLP exporter instead. Support for this exporter will end July 2023.",
+    )
     def __init__(
         self,
         collector_endpoint: Optional[str] = None,