+++ opentelemetry-proto/tests/test_proto.py
@@ -13,15 +13,12 @@
 # limitations under the License.
 # type: ignore
 
+from importlib.util import find_spec
 from unittest import TestCase
-
-from pkg_resources import DistributionNotFound, require
 
 
 class TestInstrumentor(TestCase):
     def test_proto(self):
 
-        try:
-            require(["opentelemetry-proto"])
-        except DistributionNotFound:
+        if find_spec("opentelemetry.proto") is None:
             self.fail("opentelemetry-proto not installed")