+++ opentelemetry-sdk/tests/metrics/test_periodic_exporting_metric_reader.py
@@ -12,13 +12,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import math
 from time import sleep, time_ns
-from typing import Sequence
+from typing import Optional, Sequence
 from unittest.mock import Mock
 
 from flaky import flaky
 
-from opentelemetry.sdk.metrics import Counter
+from opentelemetry.sdk.metrics import Counter, MetricsTimeoutError
 from opentelemetry.sdk.metrics._internal import _Counter
 from opentelemetry.sdk.metrics.export import (
     AggregationTemporality,
@@ -66,6 +67,25 @@ class FakeMetricsExporter(MetricExporter):
         return True
 
 
+class ExceptionAtCollectionPeriodicExportingMetricReader(
+    PeriodicExportingMetricReader
+):
+    def __init__(
+        self,
+        exporter: MetricExporter,
+        exception: Exception,
+        export_interval_millis: Optional[float] = None,
+        export_timeout_millis: Optional[float] = None,
+    ) -> None:
+        super().__init__(
+            exporter, export_interval_millis, export_timeout_millis
+        )
+        self._collect_exception = exception
+
+    def collect(self, timeout_millis: float = 10_000) -> None:
+        raise self._collect_exception
+
+
 metrics_list = [
     Metric(
         name="sum_name",
@@ -110,11 +130,13 @@ class TestPeriodicExportingMetricReader(ConcurrencyTestBase):
         pmr.shutdown()
 
     def _create_periodic_reader(
-        self, metrics, exporter, collect_wait=0, interval=60000
+        self, metrics, exporter, collect_wait=0, interval=60000, timeout=30000
     ):
 
         pmr = PeriodicExportingMetricReader(
-            exporter, export_interval_millis=interval
+            exporter,
+            export_interval_millis=interval,
+            export_timeout_millis=timeout,
         )
 
         def _collect(reader, timeout_millis):
@@ -133,6 +155,38 @@ class TestPeriodicExportingMetricReader(ConcurrencyTestBase):
         sleep(0.1)
         self.assertTrue(collect_mock.assert_called_once)
         pmr.shutdown()
+
+    def test_ticker_not_called_on_infinity(self):
+        collect_mock = Mock()
+        exporter = FakeMetricsExporter()
+        exporter.export = Mock()
+        pmr = PeriodicExportingMetricReader(
+            exporter, export_interval_millis=math.inf
+        )
+        pmr._set_collect_callback(collect_mock)
+        sleep(0.1)
+        self.assertTrue(collect_mock.assert_not_called)
+        pmr.shutdown()
+
+    def test_ticker_value_exception_on_zero(self):
+        exporter = FakeMetricsExporter()
+        exporter.export = Mock()
+        self.assertRaises(
+            ValueError,
+            PeriodicExportingMetricReader,
+            exporter,
+            export_interval_millis=0,
+        )
+
+    def test_ticker_value_exception_on_negative(self):
+        exporter = FakeMetricsExporter()
+        exporter.export = Mock()
+        self.assertRaises(
+            ValueError,
+            PeriodicExportingMetricReader,
+            exporter,
+            export_interval_millis=-100,
+        )
 
     @flaky(max_runs=3, min_passes=1)
     def test_ticker_collects_metrics(self):
@@ -186,3 +240,15 @@ class TestPeriodicExportingMetricReader(ConcurrencyTestBase):
                 self.assertTrue(isinstance(value, DefaultAggregation))
             else:
                 self.assertTrue(isinstance(value, LastValueAggregation))
+
+    def test_metric_timeout_does_not_kill_worker_thread(self):
+        exporter = FakeMetricsExporter()
+        pmr = ExceptionAtCollectionPeriodicExportingMetricReader(
+            exporter,
+            MetricsTimeoutError("test timeout"),
+            export_timeout_millis=1,
+        )
+
+        sleep(0.1)
+        self.assertTrue(pmr._daemon_thread.is_alive())
+        pmr.shutdown()