+++ opentelemetry-sdk/tests/test_configurator.py
@@ -14,11 +14,13 @@
 # type: ignore
 # pylint: skip-file
 
-import logging
+from logging import getLogger
 from os import environ
 from typing import Dict, Iterable, Optional, Sequence
 from unittest import TestCase
 from unittest.mock import patch
+
+from pytest import raises
 
 from opentelemetry import trace
 from opentelemetry.context import Context
@@ -30,6 +32,7 @@ from opentelemetry.sdk._configuration import (
     _get_exporter_names,
     _get_id_generator,
     _get_sampler,
+    _import_config_components,
     _import_exporters,
     _import_id_generator,
     _import_sampler,
@@ -306,10 +309,16 @@ class TestTraceInit(TestCase):
         environ, {"OTEL_RESOURCE_ATTRIBUTES": "service.name=my-test-service"}
     )
     def test_trace_init_default(self):
+
+        auto_resource = Resource.create(
+            {
+                "telemetry.auto.version": "test-version",
+            }
+        )
         _init_tracing(
             {"zipkin": Exporter},
             id_generator=RandomIdGenerator(),
-            auto_instrumentation_version="test-version",
+            resource=auto_resource,
         )
 
         self.assertEqual(self.set_provider_mock.call_count, 1)
@@ -349,13 +358,14 @@ class TestTraceInit(TestCase):
 
     @patch.dict(environ, {OTEL_PYTHON_ID_GENERATOR: "custom_id_generator"})
     @patch("opentelemetry.sdk._configuration.IdGenerator", new=IdGenerator)
-    @patch("opentelemetry.sdk._configuration.iter_entry_points")
-    def test_trace_init_custom_id_generator(self, mock_iter_entry_points):
-        mock_iter_entry_points.configure_mock(
+    @patch("opentelemetry.sdk._configuration.entry_points")
+    def test_trace_init_custom_id_generator(self, mock_entry_points):
+        mock_entry_points.configure_mock(
             return_value=[
                 IterEntryPoint("custom_id_generator", CustomIdGenerator)
             ]
         )
+
         id_generator_name = _get_id_generator()
         id_generator = _import_id_generator(id_generator_name)
         _init_tracing({}, id_generator=id_generator)
@@ -372,10 +382,10 @@ class TestTraceInit(TestCase):
         provider = self.set_provider_mock.call_args[0][0]
         self.assertIsNone(provider.sampler)
 
-    @patch("opentelemetry.sdk._configuration.iter_entry_points")
+    @patch("opentelemetry.sdk._configuration.entry_points")
     @patch.dict("os.environ", {OTEL_TRACES_SAMPLER: "custom_sampler_factory"})
-    def test_trace_init_custom_sampler_with_env(self, mock_iter_entry_points):
-        mock_iter_entry_points.configure_mock(
+    def test_trace_init_custom_sampler_with_env(self, mock_entry_points):
+        mock_entry_points.configure_mock(
             return_value=[
                 IterEntryPoint(
                     "custom_sampler_factory",
@@ -383,18 +393,19 @@ class TestTraceInit(TestCase):
                 )
             ]
         )
+
         sampler_name = _get_sampler()
         sampler = _import_sampler(sampler_name)
         _init_tracing({}, sampler=sampler)
         provider = self.set_provider_mock.call_args[0][0]
         self.assertIsInstance(provider.sampler, CustomSampler)
 
-    @patch("opentelemetry.sdk._configuration.iter_entry_points")
+    @patch("opentelemetry.sdk._configuration.entry_points")
     @patch.dict("os.environ", {OTEL_TRACES_SAMPLER: "custom_sampler_factory"})
     def test_trace_init_custom_sampler_with_env_bad_factory(
-        self, mock_iter_entry_points
+        self, mock_entry_points
     ):
-        mock_iter_entry_points.configure_mock(
+        mock_entry_points.configure_mock(
             return_value=[
                 IterEntryPoint(
                     "custom_sampler_factory",
@@ -402,13 +413,14 @@ class TestTraceInit(TestCase):
                 )
             ]
         )
+
         sampler_name = _get_sampler()
         sampler = _import_sampler(sampler_name)
         _init_tracing({}, sampler=sampler)
         provider = self.set_provider_mock.call_args[0][0]
         self.assertIsNone(provider.sampler)
 
-    @patch("opentelemetry.sdk._configuration.iter_entry_points")
+    @patch("opentelemetry.sdk._configuration.entry_points")
     @patch.dict(
         "os.environ",
         {
@@ -417,9 +429,9 @@ class TestTraceInit(TestCase):
         },
     )
     def test_trace_init_custom_sampler_with_env_unused_arg(
-        self, mock_iter_entry_points
+        self, mock_entry_points
     ):
-        mock_iter_entry_points.configure_mock(
+        mock_entry_points.configure_mock(
             return_value=[
                 IterEntryPoint(
                     "custom_sampler_factory",
@@ -427,13 +439,14 @@ class TestTraceInit(TestCase):
                 )
             ]
         )
+
         sampler_name = _get_sampler()
         sampler = _import_sampler(sampler_name)
         _init_tracing({}, sampler=sampler)
         provider = self.set_provider_mock.call_args[0][0]
         self.assertIsInstance(provider.sampler, CustomSampler)
 
-    @patch("opentelemetry.sdk._configuration.iter_entry_points")
+    @patch("opentelemetry.sdk._configuration.entry_points")
     @patch.dict(
         "os.environ",
         {
@@ -441,10 +454,8 @@ class TestTraceInit(TestCase):
             OTEL_TRACES_SAMPLER_ARG: "0.5",
         },
     )
-    def test_trace_init_custom_ratio_sampler_with_env(
-        self, mock_iter_entry_points
-    ):
-        mock_iter_entry_points.configure_mock(
+    def test_trace_init_custom_ratio_sampler_with_env(self, mock_entry_points):
+        mock_entry_points.configure_mock(
             return_value=[
                 IterEntryPoint(
                     "custom_ratio_sampler_factory",
@@ -452,6 +463,7 @@ class TestTraceInit(TestCase):
                 )
             ]
         )
+
         sampler_name = _get_sampler()
         sampler = _import_sampler(sampler_name)
         _init_tracing({}, sampler=sampler)
@@ -459,7 +471,7 @@ class TestTraceInit(TestCase):
         self.assertIsInstance(provider.sampler, CustomRatioSampler)
         self.assertEqual(provider.sampler.ratio, 0.5)
 
-    @patch("opentelemetry.sdk._configuration.iter_entry_points")
+    @patch("opentelemetry.sdk._configuration.entry_points")
     @patch.dict(
         "os.environ",
         {
@@ -468,9 +480,9 @@ class TestTraceInit(TestCase):
         },
     )
     def test_trace_init_custom_ratio_sampler_with_env_bad_arg(
-        self, mock_iter_entry_points
+        self, mock_entry_points
     ):
-        mock_iter_entry_points.configure_mock(
+        mock_entry_points.configure_mock(
             return_value=[
                 IterEntryPoint(
                     "custom_ratio_sampler_factory",
@@ -478,13 +490,14 @@ class TestTraceInit(TestCase):
                 )
             ]
         )
+
         sampler_name = _get_sampler()
         sampler = _import_sampler(sampler_name)
         _init_tracing({}, sampler=sampler)
         provider = self.set_provider_mock.call_args[0][0]
         self.assertIsNone(provider.sampler)
 
-    @patch("opentelemetry.sdk._configuration.iter_entry_points")
+    @patch("opentelemetry.sdk._configuration.entry_points")
     @patch.dict(
         "os.environ",
         {
@@ -492,9 +505,9 @@ class TestTraceInit(TestCase):
         },
     )
     def test_trace_init_custom_ratio_sampler_with_env_missing_arg(
-        self, mock_iter_entry_points
+        self, mock_entry_points
     ):
-        mock_iter_entry_points.configure_mock(
+        mock_entry_points.configure_mock(
             return_value=[
                 IterEntryPoint(
                     "custom_ratio_sampler_factory",
@@ -502,13 +515,14 @@ class TestTraceInit(TestCase):
                 )
             ]
         )
+
         sampler_name = _get_sampler()
         sampler = _import_sampler(sampler_name)
         _init_tracing({}, sampler=sampler)
         provider = self.set_provider_mock.call_args[0][0]
         self.assertIsNone(provider.sampler)
 
-    @patch("opentelemetry.sdk._configuration.iter_entry_points")
+    @patch("opentelemetry.sdk._configuration.entry_points")
     @patch.dict(
         "os.environ",
         {
@@ -517,24 +531,17 @@ class TestTraceInit(TestCase):
         },
     )
     def test_trace_init_custom_ratio_sampler_with_env_multiple_entry_points(
-        self, mock_iter_entry_points
+        self, mock_entry_points
     ):
-        mock_iter_entry_points.configure_mock(
+        mock_entry_points.configure_mock(
             return_value=[
-                IterEntryPoint(
-                    "custom_ratio_sampler_factory",
-                    CustomSamplerFactory.get_custom_ratio_sampler,
-                ),
                 IterEntryPoint(
                     "custom_sampler_factory",
                     CustomSamplerFactory.get_custom_sampler,
                 ),
-                IterEntryPoint(
-                    "custom_z_sampler_factory",
-                    CustomSamplerFactory.empty_get_custom_sampler,
-                ),
             ]
         )
+
         sampler_name = _get_sampler()
         sampler = _import_sampler(sampler_name)
         _init_tracing({}, sampler=sampler)
@@ -569,7 +576,7 @@ class TestLoggingInit(TestCase):
         self.processor_patch.stop()
         self.set_provider_patch.stop()
         self.provider_patch.stop()
-        root_logger = logging.getLogger("root")
+        root_logger = getLogger("root")
         root_logger.handlers = [
             handler
             for handler in root_logger.handlers
@@ -577,7 +584,12 @@ class TestLoggingInit(TestCase):
         ]
 
     def test_logging_init_empty(self):
-        _init_logging({}, "auto-version")
+        auto_resource = Resource.create(
+            {
+                "telemetry.auto.version": "auto-version",
+            }
+        )
+        _init_logging({}, resource=auto_resource)
         self.assertEqual(self.set_provider_mock.call_count, 1)
         provider = self.set_provider_mock.call_args[0][0]
         self.assertIsInstance(provider, DummyLoggerProvider)
@@ -592,7 +604,8 @@ class TestLoggingInit(TestCase):
         {"OTEL_RESOURCE_ATTRIBUTES": "service.name=otlp-service"},
     )
     def test_logging_init_exporter(self):
-        _init_logging({"otlp": DummyOTLPLogExporter})
+        resource = Resource.create({})
+        _init_logging({"otlp": DummyOTLPLogExporter}, resource=resource)
         self.assertEqual(self.set_provider_mock.call_count, 1)
         provider = self.set_provider_mock.call_args[0][0]
         self.assertIsInstance(provider, DummyLoggerProvider)
@@ -605,7 +618,7 @@ class TestLoggingInit(TestCase):
         self.assertIsInstance(
             provider.processor.exporter, DummyOTLPLogExporter
         )
-        logging.getLogger(__name__).error("hello")
+        getLogger(__name__).error("hello")
         self.assertTrue(provider.processor.exporter.export_called)
 
     @patch.dict(
@@ -633,6 +646,34 @@ class TestLoggingInit(TestCase):
         self.assertEqual(logging_mock.call_count, 1)
         self.assertEqual(tracing_mock.call_count, 1)
 
+    @patch.dict(
+        environ,
+        {
+            "OTEL_RESOURCE_ATTRIBUTES": "service.name=otlp-service",
+            "OTEL_PYTHON_LOGGING_AUTO_INSTRUMENTATION_ENABLED": "True",
+        },
+    )
+    @patch("opentelemetry.sdk._configuration._init_tracing")
+    @patch("opentelemetry.sdk._configuration._init_logging")
+    @patch("opentelemetry.sdk._configuration._init_metrics")
+    def test_initialize_components_resource(
+        self, metrics_mock, logging_mock, tracing_mock
+    ):
+        _initialize_components("auto-version")
+        self.assertEqual(logging_mock.call_count, 1)
+        self.assertEqual(tracing_mock.call_count, 1)
+        self.assertEqual(metrics_mock.call_count, 1)
+
+        _, args, _ = logging_mock.mock_calls[0]
+        logging_resource = args[1]
+        _, _, kwargs = tracing_mock.mock_calls[0]
+        tracing_resource = kwargs["resource"]
+        _, args, _ = metrics_mock.mock_calls[0]
+        metrics_resource = args[1]
+        self.assertEqual(logging_resource, tracing_resource)
+        self.assertEqual(logging_resource, metrics_resource)
+        self.assertEqual(tracing_resource, metrics_resource)
+
 
 class TestMetricsInit(TestCase):
     def setUp(self):
@@ -658,7 +699,12 @@ class TestMetricsInit(TestCase):
         self.provider_patch.stop()
 
     def test_metrics_init_empty(self):
-        _init_metrics({}, "auto-version")
+        auto_resource = Resource.create(
+            {
+                "telemetry.auto.version": "auto-version",
+            }
+        )
+        _init_metrics({}, resource=auto_resource)
         self.assertEqual(self.set_provider_mock.call_count, 1)
         provider = self.set_provider_mock.call_args[0][0]
         self.assertIsInstance(provider, DummyMeterProvider)
@@ -675,7 +721,8 @@ class TestMetricsInit(TestCase):
         {"OTEL_RESOURCE_ATTRIBUTES": "service.name=otlp-service"},
     )
     def test_metrics_init_exporter(self):
-        _init_metrics({"otlp": DummyOTLPMetricExporter})
+        resource = Resource.create({})
+        _init_metrics({"otlp": DummyOTLPMetricExporter}, resource=resource)
         self.assertEqual(self.set_provider_mock.call_count, 1)
         provider = self.set_provider_mock.call_args[0][0]
         self.assertIsInstance(provider, DummyMeterProvider)
@@ -781,4 +828,35 @@ class TestImportExporters(TestCase):
         self.assertEqual(
             metric_exporterts["console"].__class__,
             ConsoleMetricExporter.__class__,
+        )
+
+
+class TestImportConfigComponents(TestCase):
+    @patch(
+        "opentelemetry.sdk._configuration.entry_points",
+        **{"side_effect": KeyError},
+    )
+    def test__import_config_components_missing_entry_point(
+        self, mock_entry_points
+    ):
+
+        with raises(RuntimeError) as error:
+            _import_config_components(["a", "b", "c"], "name")
+        self.assertEqual(
+            str(error.value), "Requested entry point 'name' not found"
+        )
+
+    @patch(
+        "opentelemetry.sdk._configuration.entry_points",
+        **{"side_effect": StopIteration},
+    )
+    def test__import_config_components_missing_component(
+        self, mock_entry_points
+    ):
+
+        with raises(RuntimeError) as error:
+            _import_config_components(["a", "b", "c"], "name")
+        self.assertEqual(
+            str(error.value),
+            "Requested component 'a' not found in entry point 'name'",
         )