+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+- PeriodicExportingMetricReader will continue if collection times out 
+  ([#3100](https://github.com/open-telemetry/opentelemetry-python/pull/3100))
+- Fix formatting of ConsoleMetricExporter.
+  ([#3197](https://github.com/open-telemetry/opentelemetry-python/pull/3197))
+
+## Version 1.16.0/0.37b0 (2023-02-17)
+- Change ``__all__`` to be statically defined.
+  ([#3143](https://github.com/open-telemetry/opentelemetry-python/pull/3143))
 - Remove the ability to set a global metric prefix for Prometheus exporter
   ([#3137](https://github.com/open-telemetry/opentelemetry-python/pull/3137))
 - Adds environment variables for log exporter
@@ -18,9 +26,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#3128](https://github.com/open-telemetry/opentelemetry-python/pull/3128))
 - Fix validation of baggage values
   ([#3058](https://github.com/open-telemetry/opentelemetry-python/pull/3058))
+- Fix capitalization of baggage keys
+  ([#3151](https://github.com/open-telemetry/opentelemetry-python/pull/3151))
+- Bump min required api version for OTLP exporters
+  ([#3156](https://github.com/open-telemetry/opentelemetry-python/pull/3156))
+- deprecate jaeger exporters
+  ([#3158](https://github.com/open-telemetry/opentelemetry-python/pull/3158))
+- Create a single resource instance
+  ([#3118](https://github.com/open-telemetry/opentelemetry-python/pull/3118))
+
 
 ## Version 1.15.0/0.36b0 (2022-12-09)
 
+- PeriodicExportingMetricsReader with +Inf interval
+  to support explicit metric collection
+  ([#3059](https://github.com/open-telemetry/opentelemetry-python/pull/3059))
 - Regenerate opentelemetry-proto to be compatible with protobuf 3 and 4
   ([#3070](https://github.com/open-telemetry/opentelemetry-python/pull/3070))
 - Rename parse_headers to parse_env_headers and improve error message